Refactoring the log trip button
